Feat: RecordDetail.vue -> 사용자 작성 기록 및 추가 버튼 features/ui 로 기능 분리
